test/testutil: refactor class `TypicalPersons` to `TypicalModules`
We are in the first refactoring phase of morphing `persons`-oriented
Address Book application to a `modules`-oriented Degree Planner
application.

In this first phase of refactoring, let's refactor class
`TypicalPersons` to `TypicalModules` and cascade the changes to all
affected classes.
