+++ src/test/java/seedu/address/model/VersionedAddressBookTest.java
@@ -4,9 +4,9 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static seedu.address.testutil.TypicalPersons.AMY;
-import static seedu.address.testutil.TypicalPersons.BOB;
-import static seedu.address.testutil.TypicalPersons.CARL;
+import static seedu.address.testutil.TypicalModules.AMY;
+import static seedu.address.testutil.TypicalModules.BOB;
+import static seedu.address.testutil.TypicalModules.CARL;
 
 import java.util.Arrays;
 import java.util.Collections;
