+++ src/test/java/seedu/address/storage/JsonAddressBookStorageTest.java
@@ -2,10 +2,10 @@ package seedu.address.storage;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
-import static seedu.address.testutil.TypicalPersons.ALICE;
-import static seedu.address.testutil.TypicalPersons.HOON;
-import static seedu.address.testutil.TypicalPersons.IDA;
-import static seedu.address.testutil.TypicalPersons.getTypicalAddressBook;
+import static seedu.address.testutil.TypicalModules.ALICE;
+import static seedu.address.testutil.TypicalModules.HOON;
+import static seedu.address.testutil.TypicalModules.IDA;
+import static seedu.address.testutil.TypicalModules.getTypicalAddressBook;
 
 import java.io.IOException;
 import java.nio.file.Path;
