+++ src/test/java/seedu/address/ui/ModuleListPanelTest.java
@@ -4,7 +4,7 @@ import static java.time.Duration.ofMillis;
 import static org.junit.Assert.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTimeoutPreemptively;
 import static seedu.address.testutil.TypicalIndexes.INDEX_SECOND_MODULE;
-import static seedu.address.testutil.TypicalPersons.getTypicalPersons;
+import static seedu.address.testutil.TypicalModules.getTypicalModules;
 import static seedu.address.ui.testutil.GuiTestAssert.assertCardDisplaysPerson;
 import static seedu.address.ui.testutil.GuiTestAssert.assertCardEquals;
 
@@ -25,7 +25,7 @@ import seedu.address.model.module.Phone;
 
 public class ModuleListPanelTest extends GuiUnitTest {
     private static final ObservableList<Module> TYPICAL_MODULES =
-            FXCollections.observableList(getTypicalPersons());
+            FXCollections.observableList(getTypicalModules());
 
     private static final long CARD_CREATION_AND_DELETION_TIMEOUT = 2500;
