+++ src/test/java/systemtests/ClearCommandSystemTest.java
@@ -1,7 +1,7 @@
 package systemtests;
 
 import static seedu.address.commons.core.Messages.MESSAGE_UNKNOWN_COMMAND;
-import static seedu.address.testutil.TypicalPersons.KEYWORD_MATCHING_MEIER;
+import static seedu.address.testutil.TypicalModules.KEYWORD_MATCHING_MEIER;
 
 import org.junit.Test;
